chore: add cascades to non-test db
