+++ src/utilities/generatePreviewPath.ts
@@ -2,7 +2,6 @@ import { CollectionSlug } from 'payload'
 
 const collectionPrefixMap: Partial<Record<CollectionSlug, string>> = {
   posts: '/posts',
-  pages: '',
 }
 
 type Props = {
