+++ test/utils/init.ts
@@ -2,6 +2,7 @@ import { getPayload, buildConfig } from 'payload';
 import { postgresAdapter } from '@payloadcms/db-postgres'
 import { config } from '@payload-config'
 import { v4 as uuid } from 'uuid';
+import { afterSchemaInit } from '@/utilities/cascade';
 
 const initOptions = {
     secret: uuid(),
@@ -11,21 +12,7 @@ const initOptions = {
         },
         // TODO: remove once this is resolved
         // https://github.com/payloadcms/payload/issues/11177
-        afterSchemaInit: [
-            ({ schema }) => {
-              const relations = ['relations_users_sites']
-              relations.forEach((relation) => {
-                const index = Symbol.for(`drizzle:PgInlineForeignKeys`)
-                const fkeys = schema.relations[relation].table[index]
-                // Loop through the foreign keys and modify them
-                fkeys.forEach((foreignKey) => {
-                  foreignKey.onDelete = 'CASCADE'
-                  foreignKey.onUpdate = 'CASCADE'
-                })
-              })
-              return schema
-            },
-          ],
+        afterSchemaInit
       }),
 }
